requirements.txt
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cloudscraper==1.2.71
h11==0.16.0
idna==3.11
numpy==2.4.4
outcome==1.3.0.post0
pandas==3.0.2
pycparser==3.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
requests-toolbelt==1.0.0
selenium==4.43.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
